+++ config/auth.php
@@ -17,9 +17,8 @@ return [
 
     'providers' => [
         'users' => [
-            'driver' => 'eloquent-webauthn',
+            'driver' => 'eloquent',
             'model' => User::class,
-            'password_fallback' => true,
         ],
     ],