+++ database/migrations/2026_03_24_155852_create_webauthn_credentials.php
@@ -1,10 +1,35 @@
 <?php
 
+use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
-use Laragear\WebAuthn\Models\WebAuthnCredential;
+use Illuminate\Support\Facades\Schema;
 
-return WebAuthnCredential::migration()->with(function (Blueprint $table) {
-    // Here you can add custom columns to the Two Factor table.
-    //
-    // $table->string('alias')->nullable();
-});
+return new class extends Migration
+{
+    public function up(): void
+    {
+        Schema::create('webauthn_credentials', function (Blueprint $table): void {
+            $table->string('id', 510)->primary();
+            $table->string('authenticatable_type');
+            $table->unsignedBigInteger('authenticatable_id');
+            $table->uuid('user_id')->nullable();
+            $table->string('alias')->nullable();
+            $table->unsignedBigInteger('counter')->nullable();
+            $table->string('rp_id');
+            $table->string('origin');
+            $table->json('transports')->nullable();
+            $table->uuid('aaguid')->nullable();
+            $table->text('public_key');
+            $table->string('attestation_format');
+            $table->json('certificates')->nullable();
+            $table->timestamp('disabled_at')->nullable();
+            $table->timestamps();
+            $table->index(['authenticatable_type', 'authenticatable_id'], 'webauthn_user_index');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::dropIfExists('webauthn_credentials');
+    }
+};